added signup.js
